move lifo in core
