+++ kalliope/core/LIFOBuffer.py
